add seed data to product model and change avg_rating from integer to float

+++ tech_reviewer_backend/db/migrate/20210205050856_create_products.rb
@@ -4,7 +4,7 @@ class CreateProducts < ActiveRecord::Migration[6.1]
       t.string :name
       t.string :brand
       t.integer :price
-      t.integer :avg_rating
+      t.float :avg_rating
       t.string :image_url
 
       t.timestamps
